Random Forest: Fix filtered feature related bug
JIRA: MADLIB-928

Random forest filters out a feature if it has the same value for every row. If grouping is enabled this filer is applied per group. However _get_bins_grps function did not consider a case where different groups have different features. The commit fixes this issue as well as a typo in one of the related functions.