SVM: Novelty detection using 1-class SVM
Jira: MADLIB-990

Additional author: Nandish Jayaram <njayaram@pivotal.io>

In this implementation of a one-class SVM, we are piggy-backing on the existing
SVM classification. The input table to a one-class SVM does not require a
dependent variable. A maximum-margin classifier is learned that separates all
the data from the origin. The default kernel for one-class is Gaussian (rbf).

Closes #48