SVM: Add class weights for use with unbalanced data
JIRA: MADLIB-998

Added 'class_weight' in the 'params' argument. It can either be a string or a
dictionary-like mapping. In case of a string, we currently only accept
'balanced' as an option. For a mapping, the user can map values of the
dependent variable to specific double precision weights. Since, SVM only
supports binary classification at present, the class_weight mapping can only
take upto two arguments.

As part of this work, we add a 'tuple_weight' argument to the SVM aggregate.
This allows future addition of sample weights (which would be multiplied with
the class weight).