Feature: Sessionize function JIRA: MADLIB-909
This contains the implementation of the phase 1 sessionize function.
The current input parameters are the input and output table names,
along with a partition expression, the time_stamp column name and
the max_time period to consider for sessionization. The implementation
uses the window function to perform sessionization.
This commit also contains the install check file for sessionization.

This incorporates various comments received on the first and second commits.
The changes made are:
- Formatting changes in python file based on comments
- Formatting of user docs
- time_out is renamed to max_time
- max_time is of type INTERVAL and not VARCHAR anymore
- the window function query itself is changed to a cleaner version