Feature: Sessionize funtion - Phase 2
JIRA: MADLIB-1001

Based on the commentss by @fmcquillan99, @decibel and @iyerr3, the
requirements for output_cols has changed a bit, which has made the
implementation significantly simpler!
output_cols is expected to be a valid SELECT expression now. So we
are now using output_cols as is, except for replacing '*' with the
actual column names from the source table.