Sessionize function: Handling NULL for partition_expr
JIRA: MADLIB-1001

Changes here now set partition_expr to partition across the whole table
if it is originally set to NULL or '' by the user. This is in line with
how path handles partition_expr. Additionally, if output_expr has NULL
or '' as its value, it is set to '*', which is its default value. Finally,
the create_view param is also set to its default value True if NULL.
This commit also contains some formatting changes to the online help.