Elastic net: Add cross validation
JIRA: MADLIB-996

This commit adds cross validation feature to the main elastic_net
function. The CV parameters are added to 'optimizer_params' to ensure
backwards compatibility. This could potentially be moved to its own
parameter in a future version.

There is a special case in CV for elastic net: there is an ability to
compute the grid for lambda values automatically. The same procedure as
warmup lambdas is used (described in the original glmnet function in
R) to compute the lambda values.