DT: Include NULL rows in count for termination check
When the primary split feature for a node is computed, the statistics of
rows going to the true and false side don't include the rows that have
NULL value for this split feature. These "NULL" rows can only be
included in the statistics during the next pass when surrogates have
been trained. This commit ensures that in the presence of NULL rows, we
don't terminate prematurely by comparing with a lower count.