DT: Correctly encode unseen categorical features
Changes applied in commit a2f4740 added an option to treat NULL values
as a new category. This was applied by changing the encoding process of
categorical features to add a new value at the end of the list of
values. The intention with the commit was to treat new unseen, non-null
values equivalent to NULL. The encoding process, however, still encoded
the unseen categorical value as -1, which is interpreted as NULL in
underlying functions. This commit updates this process to correctly use
the last index as the encoding for the unseen/NULL value.