Update hub calculation and normalization
This commits did the following stuff together:
1. Re-wrote hub calculation. Previously, hub score was calculated by
adding authority scores of nodes that point to current node,
however, it should be the other way. Hub now is calculated by adding
authority scores of nodes that the current node points to. There are
also some corresponding changes in test file and documentation for this.
2. Changed the order of normalizing scores. Previously, for each
iteration, after authority scores are calculated, a normalization will
be done for those scores, and then hub scores will be calculated and
normalized. In this time, normalization will happen after both authority
and hub scores are calculated. Algebraically, those two approaches are
the same, so this change won't impact the final result. It just followed
the way that is suggested in the original paper
(https://www.cs.cornell.edu/home/kleinber/auth.pdf)
3. Dealt with indentation in multiple files.
4. Corrected some spelling error.