Move build dir to /madlib/docker_build
Previously, in the docker container, build dir is created in /build, but
it is not mounted to user's local machine. In this commit, build dir is
moved to /madlib/docker_build, so that it mounts to user's local madlib
repo. In this way, there is no need to transfer build artifacts and docs
between docker container and local machine.