Rename 'row_count' to 'num_col_summarized' in summary() return result
Previously, when we run `SELECT * FROM madlib.summary(valid_inputs)`, it returns a
composite type containing a filed named `row_count`, which refers to the
number of rows in the output table.
when we run `SELECT * FROM user_defined_summary_table;`, it also
contains a column named `row_count`, which refers to number of rows for
the target column.

To eliminate the confusion, we rename the first `row_count` to
`num_col_summarized`, modify explanation and also update user doc.