RF: Output non-negative importance values
Variable importance is computed in RF as the difference in prediction
accuracy between original data and permuted data from out-of-bag
samples (OOB). Permuted data is defined as each variable resampled from
its own distribution. This value can end up being negative if the number
of levels for a variable is small and is unbalanced, as the
redistribution doesn't change the data much. This commit shifts all the
importance values if some of them are negative to ensure that the lowest
importance value is 0.

Closes #231