DT: Ensure n_folds and null_proxy are set correctly
The summary table in Decision Tree included two entries: k and
null_proxy. The 'k' value is supposed to reflect the 'n_folds' value but
was not set correctly and was always NULL. The null_proxy was supposed
to be NULL when null_as_category=False but was set to the string 'NULL'.
Both these issues have been corrected in this commit.

Closes #236