DT: Add null_proxy in cross-validation
The null_proxy parameter was being reset in _build_tree, a sub-function
that is called by cross validation. This commit fixes the problem by
adding an additional parameter in the function.

Further, a scenario where no features exist due categorical columns with
single level being dropped led to the database crashing. This is now
caught and an error is thrown instead of the crash.

Closes #236