Fix boolean cast issue in gpdb4.3
In gpdb43, it is not allowed to directly cast boolean and logical
expression to text, in stead, we call madlib.__to_char() to convert it
to char first and then cast to text. This commit fix this issue in one
query so that it works for both gpdb43 and gpdb5.