Add new class for mean and std dev calculation.
Some of the modules need to calculate the mean and std dev of indep/dep
variable and this code is duplicated in all these modules.
This new class takes out the common code among all these modules and
creates a function for it.
A future commit should use this class for calculating the mean and std
dev.