Pagerank: Add option for personalized vertices
JIRA: MADLIB-1084

This commit introduces personalized Page Rank by adding a new optional
intput parameter in pagerank interface.

Personalization vertices from input parameter will have a higher jump
probability as compared to other vertices and random surfer is more
likely to jump on these personalization vertices. These personalization
vertices are initialized with an initial probabilty of 1/N where N is
the total number of vertices in the graph and rest of the vertices in
the graph are assigned an initial probability of 0. Pagerank calculated
for these vertices is biased as a random jump probability is assigned to
only these vertices during the pagerank calculation, which is equal to 1
- damping factor.

Also updated:
1. Group support
2. Install check test cases
3. Design doc

Co-authored-by: Jingyi Mei <jmei@pivotal.io>