SVM: Revert minibatch-related work
This commit is a partial revert of a8bbe08.

Minibatch was not found to be useful for SVM and broken due to recent
changes. We're disabling and reverting that work, with the task of
revisiting this again when the minibatch interface is stable.

There were some changes in a8bbe08 that were unrelated to minibatch and
useful in general context. These changes have not been reverted.

Closes #247