PCA: Fix issue with text grouping col input
JIRA: MADLIB-1215

PCA fails when the grouping column is a text column (a common use case).
This is because the column is compared to its values in a where
clause with the value not quoted. This commit adds single quotes around
the value.

Other changes include whitespace cleanup and PEP8 conforming changes.

Closes #242