DT: Fix the group value sorting in cat levels
JIRA: MADLIB-1218

In decision tree, categorical levels in the data is computed and then
sorted into multiple variables, presumably to avoid passing a dictionary
from Python to C++. This sorting was broken for grouping due to an
incorrect ordering of for-loops. This commit fixes that error and adds
multiple comments to elucidate that process.

Closes: #248