MLP: Simplify initialization of model coefficients
Changes:

1. Model initialization now happens in the C++ code instead of being
passed via Python.
2. Warm start coefficients are still passed from Python. These
coefficients are copied into the model. The vectorized form of the copy
raises an Eigen assertion when used with a MappedMatrix, possibly due
to the actual block size not being available via the Map.
3. The distinction between a "TaskState" and an "AlgoState" within a
State type has been removed in the MLPMinibatchState. This demarcation
seemed to be confusing and didn't provide much abstraction.
4. A bool named 'warm_start' was not being used and has been removed
from the main mlp aggregates (igd and minibatch).

Closes #251