MLP: Remove minibatch training dependency on original source
The original source table used in minibatch preprocessor is stored in a
column named 'original_source_table' in the summary table. Ideally, this
table doesn't need to exist after the preprocessed table is created. The
current MLP training code had a dependency on it and would fail if that
table was deleted. This commit removes that dependency.

Co-authored-by: Nikhil Kak <nkak@pivotal.io>