MLP: Remove source table dependency for predicting regression models
JIRA: MADLIB-1223

We were getting the type of the dependent variable from the
input table to know if it was an array or not. With this commit, we now get
this information from the model summary table.
We also had to add a column in the minibatch preprocessor summary table
called `dependent_vartype` in order to get the dependent type in the mlp
code for minibatching.

Closes #255

Co-authored-by: Nikhil Kak <nkak@pivotal.io>