Minibatch Preprocessor: Create temp table for standardization.
We did a few experiments and the results proved that creating a
temp table for standardization is faster than using a subquery.
This commit now creates a temp table for the standardization.
Before this commit, we were calling the `utils_normalize_data`
function inside the main query but now we create a temp table from the
output of `utils_normalize_data` and use the table in the main query.

Closes #260