MLP: Ensure grouping_col is same as preprocessed
If mini-batch preprocessor is run with grouping, the standardization in
the output table is computed per group. This implies that MLP should
also be run with the same grouping, else the dataset used for training
would be different from the original data, hence making the training
invalid.

This commit ensures that MLP training will proceed only if the grouping
column input is same as the one used during preprocessing.

Closes #263