RF: Default optional args in predict to None
Upgrading from v1.11 to v1.14 led to an error if a user had models
trainined in the older version and then predicted using the newer
version. This was due to a newly introduced argument ('null_proxy')
missing from the model summary table. This argument is optional and can
be set to NULL/None if missing.