Madpack: fix install/reinstall not giving proper error message
JIRA: MADLIB-1248

Previously, uninstalling or reinstalling on a database that does not have MADlib already
installed fails as expected. However, the info messages do not mention
this failure and also keeps going to prepare database objects to
install. This commit fixes this issue. If there is no madlib installed
in database, madpack will show nothing found to uninstall/reinstall and
stop.