Madpack: Duplicate DROP SCHEMA due to invalid cache
JIRA: MADLIB-1014

We've intermittently noticed a "cache lookup failure" due to a
"DROP OWNED BY". This was noticed only with CV enabled, possibly due to
the excessive number of tables created by CV.

This could be related to an error on StackExchange:
https://dba.stackexchange.com/questions/173815/redshift-internalerror-cache-lookup-failed-for-relation

An excerpt from the issue: "The problem is on the Postgres DB engine
caching is something to do with the System Catalog Cache access. ... the
issue gets reproduced when DROP is used very often and the cache is not
able to retrieve data which seems to be out-of-sync. After sometime (few
seconds) the cache is in back in sync and query runs fine. The only
workaround I see for the moment is retry query when fail ..."

To get around this issue, we duplicate the "DROP CASCADE" call after a
delay with the hope that the second call will clear the schema without
an issue.