Vector to columns: added support for splitting arrays of different lengths
Added support to split arrays of different lengths in the vector_col. If
the user does not provide feature names, we pad each array to the
maximum length and split across the maximum possible number of features.
If the user does provide feature names, we truncate/pad the arrays
according to the number of features the user desires. Added dev-check
tests and removed unit test to incorporate this change.

Co-authored-by: Rahul Iyer <riyer@apache.org>