Minibatch Preprocessor for Deep learning
The minibatch preprocessor we currently have in MADlib is bloated for DL
tasks. This feature adds a simplified version of creating buffers.  It
can divide each element of the independent array by a normalizing constant
for standardization (which is 255.0 by default), and optionally shift the
dependent variables by an offset if they are a numeric type.

Closes #342

Co-authored-by: Arvind Sridhar <asridhar@pivotal.io>
Co-authored-by: Domino Valdano <dvaldano@pivotal.io>