K-NN: Add kd-tree method for approximate knn
JIRA: MADLIB-1061

This commit adds a kd-tree option to the 'knn' function. A kd-tree is
used to reduce the search space to find nearest neighbors. The method
implemented here does not produce the complete kd-tree, instead it
allows the user to specify a maximum depth for the binary tree.

Additional changes:
- Add function to clean madlib views
- Move k-nn out of 'Early Stage Development'

Closes #352

Co-authored-by: Rahul Iyer <riyer@apache.org>
Co-authored-by: Frank McQuillan <fmcquillan@pivotal.io>